+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 #### Features
 - Share link from another app
 - Preload title of the page
+- Clean searched tags after saving
+- Clean search tags field when tag added
+
 # 0.1.0 (2020.05.09)
 
 #### Features